Fix regression bug for contract query
Fixing the following error message that is thrown if a QueryResult object is getting serialized for display in the console.

File "/home/wagnerm/python3.9/lib/python3.8/json/encoder.py", line 179, in default
    raise TypeError(f'Object of type {o.__class__.__name__} '
TypeError: Object of type QueryResult is not JSON serializable

It is caused because the following line only checks for inheritance of the Object class. 
https://github.com/ElrondNetwork/elrond-sdk-erdpy/blob/main/erdpy/utils.py#L32

Without this fix every call to `erdpy contract query ` fails.